changed grid code
