+++ ramp/package.json
@@ -8,6 +8,7 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "aatt": "^1.0.1",
     "core-js": "^2.6.5",
     "serialize-javascript": "^2.1.2",
     "vue": "^2.6.10",
