+++ ramp/src/components/SymbologyStackComponent.vue
@@ -36,7 +36,7 @@
 
 <script>
 import SymbologyStackIcon from "./SymbologyStackIcon";
-import GridComponent from "./GridComponent";
+import GridComponent from "./table/grid-component";
 
 export default {
   name: "SymbologyStackComponent",
