+++ ramp/src/components/table/panel-state-manager.js
